Remove API interaction from ExonumClient and fix docs
