+++ exonum_client/client.py
@@ -12,10 +12,9 @@ from logging import getLogger
 from threading import Thread
 from urllib.parse import urlencode
 from websocket import WebSocket
-import requests
 
-from .api import ServiceApi
-from .protobuf_loader import ProtobufLoader, ProtobufProviderInterface, ProtoFile
+from .api import ServiceApi, PublicApi, PrivateApi
+from .protobuf_loader import ProtobufLoader
 from .message import ExonumMessage
 
 # pylint: disable=C0103
@@ -164,8 +163,7 @@ class Subscriber:
         return response
 
 
-# pylint: disable=too-many-public-methods
-class ExonumClient(ProtobufProviderInterface):
+class ExonumClient:
     """ExonumClient class is capable of interaction with ExonumBlockchain.
 
     All the methods that perform requests to the Exonum REST API return a requests.Response object.
@@ -178,9 +176,9 @@ class ExonumClient(ProtobufProviderInterface):
     Example usage:
 
     >>> client = ExonumClient(hostname="127.0.0.1", public_api_port=8080, private_api_port=8081)
-    >>> health_info = client.health_info().json()
+    >>> health_info = client.public_api.health_info().json()
     {'consensus_status': 'Enabled', 'connected_peers': 0}
-    >>> user_agent = client.user_agent().json()
+    >>> user_agent = client.public_api.user_agent().json()
     exonum 0.13.0-rc.2/rustc 1.37.0 (eae3437df 2019-08-13)
     """
 
@@ -204,6 +202,9 @@ class ExonumClient(ProtobufProviderInterface):
         self.public_api_port = public_api_port
         self.private_api_port = private_api_port
         self.tx_url = _TX_URL.format(self.schema, hostname, public_api_port)
+
+        self.public_api = PublicApi(hostname, public_api_port, self.schema)
+        self.private_api = PrivateApi(hostname, private_api_port, self.schema)
 
     def __repr__(self) -> str:
         """ Conversion to a string. """
@@ -293,522 +294,3 @@ class ExonumClient(ProtobufProviderInterface):
         """
         subscriber = Subscriber(self.hostname, self.public_api_port, subscription_type)
         return subscriber
-
-    def service_endpoint(self, service_name: str, sub_uri: str, private: bool = False) -> str:
-        """
-        Creates a service endpoint for a given service name and sub-uri.
-
-        Example:
-
-        >>> client.service_endpoint("supervisor", "deploy-artifact", private=True)
-        http://127.0.0.1:8081/api/services/supervisor/deploy-artifact
-
-        Parameters
-        ----------
-        service_name: str
-            Name of a service instance.
-        sub_uri: str
-            Additional part of a URL to be added to the endpoint, e.g. "some/sub/uri?parameter=value"
-        private: bool
-            Denotes if a private port should be used. Defaults to False.
-
-        Returns
-        -------
-        url: str
-            Returns a service REST API URL based on the provided parameters.
-        """
-        port = self.public_api_port if not private else self.private_api_port
-
-        service_url = _SERVICE_URL.format(self.schema, self.hostname, port, service_name)
-
-        return service_url + sub_uri
-
-    # API section.
-    # Methods below perform REST API calls to an Exonum node:
-
-    def available_services(self) -> requests.Response:
-        """
-        Gets a list of available services from Exonum.
-
-        Example:
-
-        >>> available_services = client.available_services().json()
-        >>> print(json.dumps(available_services, indent=2))
-        {
-          "artifacts": [
-            {
-              "runtime_id": 0,
-              "name": "exonum-supervisor",
-              "version": "0.13.0-rc.2"
-            }
-          ],
-          "services": [
-            {
-              "spec": {
-                "id": 0,
-                "name": "supervisor",
-                "artifact": {
-                  "runtime_id": 0,
-                  "name": "exonum-supervisor",
-                  "version": "0.13.0-rc.2"
-                }
-              },
-              "status": "Active"
-            }
-          ]
-        }
-        """
-        return _get(self._system_public_endpoint("services"))
-
-    def get_instance_id_by_name(self, name: str) -> Optional[int]:
-        """
-        Gets an ID of the service instance with the provided name.
-
-        Example:
-
-        >>> id = client.get_instance_id_by_name("cryptocurrency")
-        >>> id
-        42
-
-        Parameters
-        ----------
-        name: Name of the service.
-
-        Returns
-        -------
-        result: int
-            ID of the instance.
-
-        Raises
-        ------
-        RuntimeError
-            An error will be raised if a response code is not 200.
-        """
-        response = self.available_services()
-
-        if response.status_code != 200:
-            raise RuntimeError("Couldn't get info about available services")
-
-        available_services = response.json()
-
-        for state in available_services["services"]:
-            service = state["spec"]
-            if service["name"] == name:
-                return service["id"]
-
-        return None
-
-    def send_transaction(self, message: ExonumMessage) -> requests.Response:
-        """
-        Sends a transaction into an Exonum node via REST API.
-
-        Example:
-
-        >>> response = client.send_transaction(message)
-        >>> response.json()
-        {'tx_hash': '713de312f48fe15559c0d4f7fb3f274dfbd3893a8a80d9f4224e97248f0e314e'}
-
-        Parameters
-        ----------
-        message: ExonumMessage
-            Prepared and signed an Exonum message.
-
-        Returns
-        -------
-        result: requests.Response
-            Result of the POST request.
-            If a transaction is correct and it is accepted, it will contain a JSON with a hash of the transaction.
-        """
-        response = _post(self.tx_url, data=message.pack_into_json(), headers={"content-type": "application/json"})
-        return response
-
-    def send_transactions(self, messages: Iterable[ExonumMessage]) -> List[requests.Response]:
-        """
-        Same as send_transaction, but for any iterable object over ExonumMessage.
-
-        Parameters
-        ----------
-        messages: Iterable[ExonumMessage]
-            A sequence of messages to send.
-
-        Returns
-        -------
-        results: List[requests.Response]
-            A list of responses for each sent transaction.
-        """
-        return [self.send_transaction(message) for message in messages]
-
-    def get_block(self, height: int) -> requests.Response:
-        """
-        Gets a block at the provided height.
-
-        Example:
-
-        >>> block = client.get_block(2).json()
-        >>> print(json.dumps(block, indent=2))
-        {
-          "proposer_id": 0,
-          "height": 2,
-          "tx_count": 0,
-          "prev_hash": "e686088d5323e51c096b42126a65fff59363c740ad0d8260c6c03c2e0c40ecdd",
-          "tx_hash": "c6c0aa07f27493d2f2e5cff56c890a353a20086d6c25ec825128e12ae752b2d9",
-          "state_hash": "e552443214f22721d007f1eef03f5e4d2483c31a439043eb32cd7b1faeef354f",
-          "precommits": [
-            "0a5c2...0603"
-          ],
-          "txs": [],
-          "time": "2019-09-12T09:50:49.390408335Z"
-        }
-
-        Parameters
-        ----------
-        height: int
-            A height of the required block.
-
-        Returns
-        -------
-        block_response: requests.Response
-            Result of an API call.
-            If it is successful, a JSON representation of the block will be in the response.
-        """
-        return _get(_BLOCK_URL.format(self.schema, self.hostname, self.public_api_port), params={"height": height})
-
-    # pylint: disable=too-many-arguments
-    def get_blocks(
-        self,
-        count: int,
-        earliest: Optional[int] = None,
-        latest: Optional[int] = None,
-        add_precommits: bool = False,
-        skip_empty_blocks: bool = False,
-        add_blocks_time: bool = False,
-    ) -> requests.Response:
-        """
-        Gets a range of blocks.
-
-        Blocks will be returned in a reversed order starting from the latest to the `latest - count + 1`.
-        See the `latest` parameter description for details.
-
-        Parameters
-        ----------
-        count: int
-            Amount of blocks. Should not be greater than Exonum's parameter MAX_BLOCKS_PER_REQUEST
-        earliest: Optional[int]
-            If not provided, it is considered to be the height of the earliest block in the blockchain.
-            Otherwise, a provided value will be used.
-        latest: Optional[int]
-            If not provided, it is considered to be the height of the latest block in the blockchain.
-            Otherwise, a provided value will be used.
-        add_precommits: bool
-            If True, precommits will also be taken into account.
-        skip_empty_blocks: bool
-            If True, only non-empty blocks will be returned. By default it is False.
-        add_blocks_time: bool
-            If True, then the returned `times` field of BlockRange will contain a median time from the
-            corresponding block precommits.
-
-        Returns
-        -------
-        blocks_range_response: requests.Response
-            Result of an API call.
-            If it is successful, a JSON representation of the block range will be in the response.
-        """
-        blocks_url = _BLOCKS_URL.format(self.schema, self.hostname, self.public_api_port)
-        params: Dict[str, Union[int, str]] = dict()
-        params["count"] = count
-
-        if earliest:
-            params["earliest"] = earliest
-        if latest:
-            params["latest"] = latest
-        if add_precommits:
-            params["add_precommits"] = "true"
-        if skip_empty_blocks:
-            params["skip_empty_blocks"] = "true"
-        if add_blocks_time:
-            params["add_blocks_time"] = "true"
-
-        return _get(blocks_url, params=params)
-
-    def get_tx_info(self, tx_hash: str) -> requests.Response:
-        """
-        Gets information about the transaction with the provided hash.
-
-        Example:
-
-        >>> tx_info = client.get_tx_info(tx_hash).json()
-        >>> print(json.dumps(tx_info, indent=2))
-        {
-          'type': 'committed',
-          'content': '0a11...660d',
-          'location': {
-            'block_height': 58224,
-            'position_in_block': 1
-          },
-          'location_proof': {
-            'proof': [
-              {
-                'index': 0,
-                'height': 1,
-                'hash': '14637aa10b700cebfbc23d45395e8677d1fe1914d2e7f50d38cf1b73cfba1702'
-              }
-            ],
-            'entries': [
-              [1, 'e2d9ba5e8e104d65be8d3af7c26e5abea8f27da280cea110a80c9ab4f4d2a10c']
-            ],
-            'length': 2
-          },
-          'status': {
-            'type': 'success'
-          },
-          'time': '2019-09-12T13:08:10.528537286Z'
-        }
-
-        Parameters
-        ----------
-        tx_hash: str
-            A hexadecimal representation of the transaction hash.
-
-        Returns
-        -------
-        block_response: requests.Response
-            Result of an API call.
-            If it is successful, a JSON representation of the transaction info will be in the response.
-        """
-        return _get(_TX_URL.format(self.schema, self.hostname, self.public_api_port), params={"hash": tx_hash})
-
-    def get_service(self, service_name: str, sub_uri: str, private: bool = False) -> requests.Response:
-        """
-        Performs a GET request to the endpoint generated by the `service_endpoint` method.
-
-        Parameters are the same as in `service_endpoint`.
-
-        Returns
-        -------
-        response: requests.Response
-            Result of an API call.
-        """
-        return _get(self.service_endpoint(service_name, sub_uri, private))
-
-    def post_service(self, service_name: str, sub_uri: str, data: str, private: bool = False) -> requests.Response:
-        """
-        Performs a POST request to the endpoint generated by the `service_endpoint` method.
-
-        Parameters are the same as in `service_endpoint` except for `data`.
-        `data` is expected to be a serialized JSON value.
-
-        Returns
-        -------
-        response: requests.Response
-            Result of an API call.
-        """
-        json_headers = {"content-type": "application/json"}
-        return _post(self.service_endpoint(service_name, sub_uri, private), data=data, headers=json_headers)
-
-    def health_info(self) -> requests.Response:
-        """ Performs a GET request to the healthcheck Exonum endpoint. """
-        return _get(self._system_public_endpoint("healthcheck"))
-
-    def stats(self) -> requests.Response:
-        """ Performs a GET request to the stats Exonum endpoint. """
-        return _get(self._system_public_endpoint("stats"))
-
-    def user_agent(self) -> requests.Response:
-        """ Performs a GET request to the user_agent Exonum endpoint. """
-        return _get(self._system_public_endpoint("user_agent"))
-
-    def add_peer(self, address: str, public_key: str) -> requests.Response:
-        """
-        Performs a POST request to the '{system_base_path}/peers' endpoint in order to add a new peer to the node.
-
-        Parameters
-        ----------
-        address: IP address of the node which the present node should connect to.
-        public_key: public key of the node which the present node should connect to.
-
-        Returns
-        -------
-        response: requests.Response
-            Result of an API call.
-        """
-        data = json.dumps({"address": address, "public_key": public_key})
-        response = _post(
-            self._system_private_endpoint("peers"), data=data, headers={"content-type": "application/json"}
-        )
-        return response
-
-    def get_peers(self) -> requests.Response:
-        """
-        Performs a GET request to the '{system_base_path}/peers' to get a list of peers of the node.
-
-        Example:
-
-        >>> peers = client.get_peers().json()
-        >>> print(json.dumps(peers, indent=2))
-        {
-          "incoming_connections": [{
-            "address": "127.0.0.1:57671",
-            "public_key": "8a17bdfe42c10abdb7f27b5648691db3338400c27812e847e02eb7193ad490f2"
-          }],
-          "outgoing_connections": {
-            "127.0.0.1:6334": {
-              "public_key": "dcb46dceaeb7d0eab7b6ed000f317f2ab9f7c8423ec9a6a602d81c0979e1333a",
-              "state": {
-                "type": "Active"
-              }
-            },
-            "127.0.0.1:6335": {
-              "public_key": "dcb46dceaeb7d0eab7b6ed000f317f2ab9f7c8423ec9a6a602d81c0979e1333a",
-              "state": {
-                "delay": 4000,
-                "type": "Reconnect"
-              }
-            },
-            "127.0.0.1:6336": {
-              "public_key": null,
-              "state": {
-                "type": "Active"
-              }
-            },
-            "127.0.0.1:6337": {
-              "public_key": null,
-              "state": {
-                "delay": 4000,
-                "type": "Reconnect"
-              }
-            }
-          }
-        }
-
-        Returns
-        -------
-        response: requests.Response
-            Result of an API call.
-            If it is successful, a list of peers will be returned.
-        """
-        return _get(self._system_private_endpoint("peers"))
-
-    def set_consensus_interaction(self, enabled: bool = True) -> requests.Response:
-        """
-        Performs a POST request to the '{system_base_path}/consensus_enabled'
-        to switch consensus interaction of the node on or off.
-
-        Parameters
-        ----------
-        enabled: flag that switches consensus interaction of the node on (True) or off (False).
-
-        Returns
-        -------
-        response: requests.Response
-            Result of an API call.
-        """
-        data = json.dumps({"enabled": enabled})
-        return _post(
-            self._system_private_endpoint("consensus_enabled"), data=data, headers={"content-type": "application/json"}
-        )
-
-    def get_consensus_interaction(self) -> requests.Response:
-        """
-        Performs a GET request to the '{system_base_path}/consensus_enabled' to get boolean value that states
-        if the node participates in consensus.
-
-        Example:
-
-        >>> client.get_consensus_interaction().json()
-        True
-
-        Returns
-        -------
-        response: requests.Response
-            Result of an API call.
-        """
-        return _get(self._system_private_endpoint("consensus_enabled"))
-
-    def get_network_info(self) -> requests.Response:
-        """
-        Performs a GET request to the '{system_base_path}/network'
-        to get info about the serialization protocol and the services functioning in the network.
-
-        Example:
-
-        >>> network_info = client.get_network_info().json()
-        >>> print(json.dumps(network_info, indent=2))
-        {
-          "core_version": "0.10.2"
-        }
-
-        Returns
-        -------
-        response: requests.Response
-            Result of an API call.
-        """
-        return _get(self._system_private_endpoint("network"))
-
-    def shutdown(self) -> requests.Response:
-        """
-        Performs a POST request to the '{system_base_path}/shutdown' to stop the node.
-        After receiving this request, the node stops processing transactions, participating in consensus and
-        terminates after all messages in the event queue are processed.
-
-        Returns
-        -------
-        response: requests.Response
-            Result of an API call.
-        """
-        data = json.dumps(None)
-        return _post(self._system_private_endpoint("shutdown"), data=data, headers={"content-type": "application/json"})
-
-    # Implementation of ProtobufProviderInterface:
-    def _get_proto_sources(self, params: Optional[Dict[str, str]] = None) -> List[ProtoFile]:
-        proto_sources_endpoint = _RUST_RUNTIME_URL.format(
-            self.schema, self.hostname, self.public_api_port, "proto-sources"
-        )
-        response = _get(proto_sources_endpoint, params=params)
-        if response.status_code != 200 or "application/json" not in response.headers["content-type"]:
-            logger.critical(
-                "Unsuccessfully attempted to retrieve Protobuf sources.\n" "Status code: %s,\n" "body:\n%s",
-                response.status_code,
-                response.content,
-            )
-            raise RuntimeError("Unsuccessfully attempted to retrieve Protobuf sources: {!r}".format(response.content))
-        logger.debug("Protobuf sources retrieved successfully.")
-
-        proto_files = [
-            ProtoFile(name=proto_file["name"], content=proto_file["content"]) for proto_file in response.json()
-        ]
-
-        return proto_files
-
-    def get_main_proto_sources(self) -> List[ProtoFile]:
-        # Performs a GET request to the `proto-sources` Exonum endpoint:
-        params = {"type": "core"}
-        return self._get_proto_sources(params)
-
-    def get_proto_sources_for_artifact(
-        self, runtime_id: int, artifact_name: str, artifact_version: str
-    ) -> List[ProtoFile]:
-        # Raise an exception if runtime ID is not equal to the rust runtime ID
-        if runtime_id != RUST_RUNTIME_ID:
-            err_msg = f"Provided runtime ID: {runtime_id} is not equal to Rust runtime ID: {RUST_RUNTIME_ID}."
-            logger.critical(err_msg)
-            raise RuntimeError(err_msg)
-        # Performs a GET request to the `proto-sources` Exonum endpoint with a provided artifact name:
-        params = {"type": "artifact", "name": artifact_name, "version": artifact_version}
-
-        return self._get_proto_sources(params)
-
-    def _system_public_endpoint(self, endpoint: str) -> str:
-        return _SYSTEM_URL.format(self.schema, self.hostname, self.public_api_port, endpoint)
-
-    def _system_private_endpoint(self, endpoint: str) -> str:
-        return _SYSTEM_URL.format(self.schema, self.hostname, self.private_api_port, endpoint)
-
-
-def _get(url: str, params: Optional[Dict[Any, Any]] = None) -> requests.Response:
-    # Internal wrapper over requests.get:
-    return requests.get(url, params=params)
-
-
-def _post(url: str, data: str, headers: Dict[str, str]) -> requests.Response:
-    # Internal wrapper over requests.post:
-    return requests.post(url, data=data, headers=headers)
